fixed small bug with closing pop-up window for profile editor
